Put Snackbar on Home, not on component Fixes #37
